Introduce tabs
Also the sketch of a store
