+++ frontend/index.tsx
@@ -1,5 +1,8 @@
 import { render } from "solid-js/web";
 import { HopeProvider, NotificationsProvider } from "@hope-ui/solid";
+import { Message } from "../types/message";
+
+import { addMessage } from "./store";
 
 import App from "./App";
 
@@ -7,11 +10,9 @@ const url = `ws://${document.location.host}/webxdc-message`;
 
 const socket = new WebSocket(url);
 
-socket.addEventListener("open", () => {
-  console.log("client web socket open");
-});
 socket.addEventListener("message", (event) => {
-  console.log("from socket:", JSON.stringify(event.data));
+  // const message = JSON.parse((event as any).data) as Message;
+  // addMessage(message);
 });
 
 render(
